C code for an optimized version of Sequential Search
Implementation of optimized version of Sequential Search. In it the key element is saved at the end of list of records and then we compare each record with key element. Unlike normal version of sequential search, this technique uses 1 comparison less(i.e we don't check if i<=N). This form of sequential search is used in DBMS etc.